requirements.txt
pyautogui
requests
psutil
pycryptodome
pywin32
pyperclip